plugins/ai-tools/skills/thesys-generative-ui/templates/python-backend/requirements.txt
# TheSys Generative UI - Python Backend Dependencies

# Core dependencies
thesys-genui-sdk>=0.1.0
openai>=1.59.5
python-dotenv>=1.0.1

# FastAPI dependencies (for fastapi-chat.py)
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.10.5

# Flask dependencies (for flask-chat.py)
flask>=3.1.0
flask-cors>=5.0.0

# Optional: For enhanced error handling
python-multipart>=0.0.20
